Improved MessageBanner
Moved to layout directory and added to layout component. This way, I don't have to add in the component all over the place. Also made MessageBanner take information directly from the router object so I don't have to worry about passing parameters to it. It simply updates and shows anytime the route changes and has a message query parameter.