Sort filter and private recipes
Pipeline was missing a step for only returning the user's own recipes when in their own recipe box. That has been fixed. Sort methods have been changed to have "New" (default), "User", and "Rating" to only have useful sorting methods, and that logic has been added to the filter object and migrated to the backend as well